+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
+
 ### Added
+
 * Read osm.xml int structures
 * Provide filter functions to filter parsed osm objects
-* Initial release
+* Write nodes and ways as json